Displaying derived seasonal food
- Works for California, but there seems to be a timing issue where the
'usaState' isn't ready when the Seasonal array look-up is happening.
I should put the logic for 'deriveVeggies' inside 'deriveLocation' to
make sure the Seasonal array lookup (by usaState) happens after the
usaState is derived.

- There's also some problem with the Node server where 'bundle.js' comes
back with '404 Not Found', but maybe this error is related to the above
paragraph. I had the server working without issues before I was working
on front-end only and now when I test the Node server again, there is
a 404 error.